requirements.txt
chromadb==1.0.13
google-generativeai==0.8.5
huggingface-hub==0.33.0
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.8
sentence-transformers==4.1.0
streamlit==1.46.0
python-dotenv==1.1.0